AcneSkinCare/gateway/requirements-gateway.txt
fastapi
uvicorn
pydantic
numpy
pillow
opencv-python-headless
requests
python-multipart
streamlit==1.10.0